address issues from review- except format flag map
