--- spec/examples/util/test-translation.json
@@ -1,6 +1,0 @@
-{
-  "simple-key": "test string",
-  "key-with-an-format": "string with {{tag, en-handle-an}} {{tag}}",
-  "key-with-capitalize-format": "{{tag, capitalize}}",
-  "key-with-multiple-formats": "string with {{tag, en-handle-an|capitalize}} {{tag}}"   
-}
